+++ CHANGES.rst
@@ -89,6 +89,7 @@ Development
   ``UdpServer`` in order to shut down using a single ``CTRL-C``
 - Upgrade to MicroWebSrv2 2.0.3. Thanks, @jczic!
 - Improve thread shutdown behavior when receiving ``SIGINT`` / ``CTRL-C``
+- Improve sensor reading from 1-Wire DS18X20 devices
 
 
 2019-08-19 0.6.0